Bug fixed

+++ src/main/java/it/unipi/lsmd/controller/TripServlet.java
@@ -23,37 +23,47 @@ import java.util.HashMap;
 public class TripServlet extends HttpServlet {
 
     @Override
-    protected void doGet(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws ServletException, IOException {
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        processRequest(req,resp);
+    }
 
+    @Override
+    protected void doGet(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws ServletException, IOException {
+        processRequest(httpServletRequest,httpServletResponse);
+    }
+
+    private void processRequest(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws ServletException, IOException {
         AuthenticatedUserDTO authenticatedUserDTO = SecurityUtils.getAuthenticatedUser(httpServletRequest);
 
         TripService tripService = ServiceLocator.getTripService();
 
         String targetJSP = "/WEB-INF/pages/trip.jsp";
         String trip_id = httpServletRequest.getParameter("id");
+        TripDetailsDTO trip = null;
+        if(trip_id!=null) {
+            trip = tripService.getTrip(trip_id);
 
-        TripDetailsDTO trip = tripService.getTrip(trip_id);
-
-        if(authenticatedUserDTO != null) {
-            LocalDateTime last_update = tripService.wishlistUpdateTime(authenticatedUserDTO.getUsername(), trip.getId());
-            if(last_update == null)
-                trip.setInWishlist(false);
-            try {
-
-                String action = httpServletRequest.getParameter("action");
-
-                if((action.equals("add") && last_update == null) || (last_update != null && trip.getLast_modified().isAfter(last_update))){
-                    TripSummaryDTO tripSummary = TripUtils.tripSummaryFromTripDetails(trip);
-                    tripService.addToWishlist(authenticatedUserDTO.getUsername(), trip_id, tripSummary);
-                    trip.setInWishlist(true);
-                }else if(action.equals("remove") && last_update != null){
-                    tripService.removeFromWishlist(authenticatedUserDTO.getUsername(), trip_id);
+            if (authenticatedUserDTO != null) {
+                LocalDateTime last_update = tripService.wishlistUpdateTime(authenticatedUserDTO.getUsername(), trip.getId());
+                if (last_update == null)
                     trip.setInWishlist(false);
-                }
+                try {
 
-            } catch (NullPointerException e) {
+                    String action = httpServletRequest.getParameter("action");
+
+                    if ((action.equals("add") && last_update == null) || (last_update != null && trip.getLast_modified().isAfter(last_update))) {
+                        TripSummaryDTO tripSummary = TripUtils.tripSummaryFromTripDetails(trip);
+                        tripService.addToWishlist(authenticatedUserDTO.getUsername(), trip_id, tripSummary);
+                        trip.setInWishlist(true);
+                    } else if (action.equals("remove") && last_update != null) {
+                        tripService.removeFromWishlist(authenticatedUserDTO.getUsername(), trip_id);
+                        trip.setInWishlist(false);
+                    }
+
+                } catch (NullPointerException e) {
+                }
+                httpServletRequest.setAttribute(SecurityUtils.STATUS, tripService.getJoinStatus(trip_id, authenticatedUserDTO.getUsername()));
             }
-            httpServletRequest.setAttribute(SecurityUtils.STATUS, tripService.getJoinStatus(trip_id,authenticatedUserDTO.getUsername()));
         }
         httpServletRequest.setAttribute(SecurityUtils.TRIP, trip);
